# Conflicts:
#	src/main/java/com/example/adressbook/AdressbookApplication.java
#	src/main/java/com/example/adressbook/controller/ContactController.java
#	src/main/java/com/example/adressbook/entity/Contact.java
#	src/main/java/com/example/adressbook/service/ContactService.java

+++ src/main/java/com/example/adressbook/controller/ContactController.java
@@ -1,11 +1,13 @@
 package com.example.adressbook.controller;
 
 import com.example.adressbook.entity.Contact;
+import com.example.adressbook.repository.ContactRepository;
 import com.example.adressbook.service.ContactService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
 
 @RestController
 public class ContactController {
@@ -13,19 +15,31 @@ public class ContactController {
     @Autowired
     private ContactService contactService;
 
+    @Autowired
+    private ContactRepository contactRepository;
+
     @GetMapping(value = "/all")
+    @CrossOrigin(origins = "http://localhost:4200")
     public List<Contact> getAllContacts() {
-        return contactService.getAllContacts();
+        return contactService.getAllContact();
     }
 
-    @PostMapping(value = "/create/{name}")
-    public void addContact(@RequestParam String name) {
-        contactService.addContactByName(name);
+    @GetMapping(path = "/{id}")
+    public String getCertainContact(@PathVariable String id) {
+        return "The contact " + contactService.getContact(id).toString() + " is here";
+    }
+
+    @PostMapping(value = "/create")
+    public String addContact(@RequestBody Contact contact) {
+        Contact contact1 = contactService.addContact(contact);
+        return "Contact was added!!  " + contact1;
     }
 
     @DeleteMapping(path = "/delete/{id}")
-    public String delete(@RequestParam String id) {
-
+    public String delete(@PathVariable String id) {
+        if (!contactRepository.existsById(id)) {
+            throw new IllegalAccessError("Contact does not now exist");
+        }
         contactService.delete(id);
         return "the contact with id: " + id + " was deleted";
     }
